pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Article Analysis Web Application using Streamlit and crewAI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "crewai>=0.11.0",
    "trafilatura>=7.0.0",
    "google-generativeai>=0.3.0",
    "requests>=2.31.0",
    "sqlite3",
    "python-dotenv>=1.0.0"
]
